Add CollaboratorAccessList component
Adding this component so it can be later used in the
invitation form as well.
