+++ src/client/components/CollaboratorAccessList/CollaboratorAccessList.scss
@@ -1,0 +1,6 @@
+@import 'src/client/style/partials';
+
+.premission-list__container {
+  text-align: left;
+  margin-bottom: $spacing-xxs;
+}
