+++ src/client/components/EditUserForm/CollaboratorPermissions/CollaboratorPermissions.tsx
@@ -2,13 +2,10 @@ import './CollaboratorPermissions.scss'
 import React, { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { Objects } from 'utils/objects'
-
-import { SubSections } from 'meta/assessment'
 import { Collaborator, CollaboratorPermissions as CollabPermissions } from 'meta/user'
 
-import { useCycle } from 'client/store/assessment'
 import { useSections } from 'client/store/metadata'
+import CollaboratorAccessList from 'client/components/CollaboratorAccessList'
 import CollaboratorAccessModal from 'client/components/CollaboratorAccessModal'
 
 type Props = {
@@ -18,7 +15,6 @@ type Props = {
 const CollaboratorPermissions = (props: Props) => {
   const { userRole } = props
   const { t } = useTranslation()
-  const cycle = useCycle()
   const sections = useSections()
 
   const [modalOptions, setModalOptions] = useState<{ open: boolean }>({ open: false })
@@ -33,45 +29,14 @@ const CollaboratorPermissions = (props: Props) => {
 
   if (!sections) return null
 
-  const options = SubSections.getAnchorsByUuid({ cycle, sections })
-
   const permissions = (userRole.permissions as CollabPermissions) || undefined
-
-  const sectionPermissions = Objects.isEmpty(permissions) ? 'all' : permissions.sections ?? 'none'
-
-  const tableDataPermissions = Object.entries(sectionPermissions)
-    .reduce((prev, [k, v]) => (v.tableData ? [...prev, options[k]] : prev), [])
-    .sort()
-    .join(', ')
-
-  const descriptionsPermissions = Object.entries(sectionPermissions)
-    .reduce((prev, [k, v]) => (v.descriptions ? [...prev, options[k]] : prev), [])
-    .sort()
-    .join(', ')
 
   return (
     <div className="edit-user__form-item edit-user__form-item-permissions">
       <div className="edit-user__form-label">{t(`userManagement.permissions`)}</div>
 
       <div className="edit-user__form-field edit-user__form-field-premissions">
-        <div className="edit-user__form-field-premission-list">
-          {typeof sectionPermissions === 'string' ? (
-            t(`contactPersons.${sectionPermissions}`)
-          ) : (
-            <>
-              {tableDataPermissions !== '' && (
-                <div>
-                  {t('userManagement.permissionNames.tableData')}: {tableDataPermissions}
-                </div>
-              )}
-              {descriptionsPermissions !== '' && (
-                <div>
-                  {t('userManagement.permissionNames.descriptions')}: {descriptionsPermissions}
-                </div>
-              )}
-            </>
-          )}
-        </div>
+        <CollaboratorAccessList permissions={permissions} />
 
         <button className="btn-xs btn-primary" onClick={_onEditPermissionsClick} type="button">
           {t('userManagement.editPermissions')}
